Incorrect resize when in SplitView mode
Fixes #121753357

When in SplitView mode, if you tried to make a window smaller but
cancelled the gesture, sometimes the resize finishes. The terminal size
was still good, but the internal number of columns was unchanged. It
looks like the events might be competing themselves, so by using the
function resizeTo nstead of doing it manually, we get it to work.