Sessions with terminal mode settings
The terminal now exposes a method to change the "mode", and the apps
make use of it.
This isn't actually the right way to do it, because our terminal only
has a real raw mode, but it will help with certain scenarios where the
apps are expecting a transformation to "app mode", specially curses
ones. A better future implementation might have to deal with this at the
ssh level.