Exit session from MCP
Closes #44. When the Terminal Session finishes, normally due to an
"exit" command on the MCP, the corresponding Space gets closed too.

This has been implemented by adding a delegate to the Session that will
get notified when this event happens, allowing for other operations to
occur. Normally, the delegate will be our TerminalController, adding
another layer of abstraction to the Session.