ssh session hangs after connection
Fixes #84. When creating a ssh or mosh session, most of the time it got
stuck at the poll in the main loop. This is because, depending on the
speed of the connection, there would be a race condition with the poll
that would block it from obtaining any data. The libssh2_read function
always has to be called before blocking in any case. And that's why the
wait_socket function works in this scenarios but the poll's main loop
got stuck.