Hotfix: Perform lossy encoding when UTF-8 fails
Fixes #176

Blink is only compatible with UTF-8 at the moment. SSH does not perform
any type of format validation, so it might happen that a non-utf8 character is
sent. We try to configure the destination environment, and also perform
lossy encoding when nothing else is possible. This is actually the
behavior on terminal.app and iTerm.