Lock Blink with TouchID / Passcode
Closes #58

As Blink will handle critical data and will never disconnect, it is
important to lock the app to avoid an intruder get access to an unlocked
device and get private data from it.

We have decided to lock the app after the device has been locked
too, instead of doing it time based. This is because time based does not
add any additional security and it is only a hassle to the user. We
wanna cover the critical case when you unlock the device for someone to
use, and do not want them taking control of the app.

Implementation is done through SwiftPasscodeLock module. It takes into
account saving the passcode securely using the keychain too.