Enforce suspension when protected data will disappear
Fixes #1048
When the device is locked and the application is still active, both
Active-Background and ProtectedDataWillBecomeUnavailable will compete between
them.

When the call order is Unavailable -> Active -> Background, the state will be
saved, then restored, and then scheduled for saving again. The problem is this
scheduling will trigger a saving outside the "safe window".

This patch takes into account what the final state should be, instead of
transitioning from event to event. We add the enforceSuspension variable so that
the Background process will save immediately instead of scheduling, if it is
aware that DataWillBecomeUnavailable.

I'm keeping the traces of the events here.