COPY Closing file may crash if the translator does not account for size
The copy algorithm closes the file once all data has been written. As the whole
architecture is asynchronous, reads may still be scheduled depending on the
Translator. That was the case of SFTP, which does not know how long the file is,
and it just schedules more blocks. In that case, the copy algorithm may close
the file earlier than the writeTo algorithm finishes, and make subsequent reads
fail.

This was only happening when the file fit exactly in 20 blocks. We should
enhance tests with this corner case.