Refactored Publisher lock
The previous lock was failing in some specific cases as it was not able to
reacquire and end up in a loop, or it would fail with an unfair_lock recursive
abort.

Two things need to happen for the cancel to work. We had to signal the flow not
to continue, as the Cancels happen on different threads, and so when the Cancel
moves up it will be scheduled but not executed on the other thread running the
operation. The second is that when the Cancel is sent, we need to block it and
ensure that it does not continue going up. Otherwise Combine may unlock while we
are still awaiting for the RunLoop.

To solve this, we are introduing a simpler Atomic value that controls the stop
sequence, and a Sleep timer of roughly the same RunLoop wait time, to make sure
that the flow does not continue.