Block channel when executing SSH operations
This tracks changes on LibSSH so that blocking calls are actually blocking the
runloop, which is important for concurrent calls on SFTP.

Libssh sets the whole session to block or unblock when a channel is set. We want
everything to be non-blocking, but SFTP is structured in a way that all
functions are blocking. We wrap each of them to set the session properly after
execution.