Files.app fixes for multiple initialization issues
- When the root enumerator is requested for the first time, there is a chance
that multiple subscribers are connected to it. This may provoke multiple
initializations in the cache overwriting each other and eventually, Translators
hanging without connections. It is similar to something that happened on VS Code
when reconnecting. The solution is to use shareReplay, so multiple initial
subscribers to a translator will see the same.

- The SSH Dialer was using a future to create a thread for a specific
connection. The Files.app may request multiple Root Enumerators on start and for whatever
reason, invalidate all except one. Because the future is executed right at the
start, this would cause multiple unused threads to be hanging there. Added a
Deferred for this scenario.