RAW branch non-versioning
RAW is a branch for testing and in continuous evolution. It isn't
intented to branch from, its commits might change, be rebased,
disappear, etc... It is unstable.

From RAW we branch the next versions of Blink from all the commits that
are ready for production, and that is what we send to testers. We
stabilise that and finally go to production.

Nightly builds are taken from here and pushed for internal testing. We
drop a 999 version so that iTunes doesn't complaint.